airbnb

+++ 2-do_deploy_web_static.py
@@ -1,31 +1,30 @@
 #!/usr/bin/python3
 """
-This module supplies a fabric script that configures a server
+Fabric script based on the file 1-pack_web_static.py that distributes an
+archive to the web servers
 """
-from fabric.api import local, task, put, run, env
-from datetime import datetime
-from os.path import basename
 
-env.user = "ubuntu"
-env.hosts = ['54.236.24.158', '34.229.161.195']
-env.key_filename = '~/.ssh/id_rsa'
+from fabric.api import put, run, env
+from os.path import exists
+env.hosts = ['142.44.167.228', '144.217.246.195']
 
 
-@task
 def do_deploy(archive_path):
-    """distributes ana rchive to the servers"""
+    """distributes an archive to the web servers"""
+    if exists(archive_path) is False:
+        return False
     try:
-        dn = basename(archive_path)[:-4]
+        file_n = archive_path.split("/")[-1]
+        no_ext = file_n.split(".")[0]
+        path = "/data/web_static/releases/"
         put(archive_path, '/tmp/')
-        run(f'mkdir -p /data/web_static/releases/{dn}/')
-        run(f'tar -xzf /tmp/{dn}.tgz -C /data/web_static/releases/{dn}/')
-        run(f'rm /tmp/{dn}.tgz')
-        run(f'mv /data/web_static/releases/{dn}/web_static/* '
-            f'/data/web_static/releases/{dn}/')
-        run('rm -rf /data/web_static/releases/web_static')
+        run('mkdir -p {}{}/'.format(path, no_ext))
+        run('tar -xzf /tmp/{} -C {}{}/'.format(file_n, path, no_ext))
+        run('rm /tmp/{}'.format(file_n))
+        run('mv {0}{1}/web_static/* {0}{1}/'.format(path, no_ext))
+        run('rm -rf {}{}/web_static'.format(path, no_ext))
         run('rm -rf /data/web_static/current')
-        run(f'ln -s /data/web_static/releases/{dn}/ /data/web_static/current')
-        print("New version deployed!")
+        run('ln -s {}{}/ /data/web_static/current'.format(path, no_ext))
         return True
-    except Exception:
+    except:
         return False
